consensus/engine: rename bft facade package to engine
