+++ consensus/engine/engine.go
@@ -1,4 +1,4 @@
-package bft
+package engine
 
 import (
 	"crypto/ecdsa"
