+++ consensus/engine/sealer_test.go
@@ -1,4 +1,4 @@
-package bft
+package engine
 
 import (
 	"bytes"
